Extract HTML block factory
